pyproject.toml
[project]
name = "tir-agent"
version = "0.1.0"
description = "Tool-Integrated Reasoning Agent based on Qwen-Agent"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "qwen-agent>=0.0.5",
    "streamlit>=1.30.0",
    "python-dotenv>=1.0.0",
    "pillow>=10.0.0",
    "pypdf>=4.0.0",
    "pdf2image>=1.16.0",
    "python-docx>=1.1.0",
    "openpyxl>=3.1.0",
    "python-pptx>=0.6.23",
    "httpx>=0.27.0",
    "openai>=1.0.0",
    "pyyaml>=6.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "playwright>=1.40.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
